wallet mod describe to file
added a doc comment

moved some spend methods from spend_wallet to send and fixed a mod typo

The combination hides some not-compiling code where one of the mod names was wrong.

renamed to mod witnesses

rearranged describe.rs

squashing again to sign
